Update infinispan config to support off-heap memory and fix missing cache config

+++ bin/test-setup.sh
@@ -31,8 +31,12 @@ fi
 pushd $DIR
 
 pushd $DIR/deployments/launcher/target/
-rm -rf indy
-tar -zxvf indy-launcher-*-complete.tar.gz
+if [ "$1" == "-e" ]; then
+  echo "Use existing $DIR/deployments/launcher/target/indy"
+else
+  rm -rf indy
+  tar -zxvf indy-launcher-*-complete.tar.gz
+fi
 
 if [ "x${TEST_REPOS}" != "x" ]; then
   echo "Copying repository/group definitions from: ${TEST_REPOS}"
